use mnemonic in registrar cli
